Restructure frontend plugins
