+++ client/coral-framework/loaders/plugins-loader.js
@@ -1,38 +1,13 @@
 /**
  * Executes `source` to retrieve plugins configuration
- * and loads all modules of specified plugins.
+ * and loads the `index.js` of specified plugins.
  *
  * Outputs a module that looks like the following:
  *
- * module.exports.actions = [{plugin: string, getModule: callback}, ...]
- * module.exports.reducer = [{plugin: string, getModule: callback}, ...]
- * [...]
+ * module.exports = [{plugin: string, module: object}, ...]
  *
  */
 const {stripIndent} = require('common-tags');
-const fs = require('fs');
-const path = require('path');
-
-function moduleExists(loc) {
-  const fileExists = fs.existsSync(path.resolve(__dirname, loc));
-  if (fileExists) {
-    return true;
-  }
-
-  const hasExtension = /\.[^\/\\]*$/.test(loc);
-  if (hasExtension) {
-    return false;
-  }
-
-  // Find file with appended `.js` extension.
-  return fs.existsSync(path.resolve(__dirname, `${loc}.js`));
-}
-
-function addIfExists(list, plugin, resource) {
-  if (moduleExists(`../../../plugins/${plugin}/client/${resource}`)) {
-    list.push(`{getModule: () => require('plugins/${plugin}/client/${resource}'), plugin: '${plugin}'}`);
-  }
-}
 
 function getPluginList(config) {
   return config.client.map(x => typeof x === 'string' ? x : Object.keys(x)[0]);
@@ -43,34 +18,14 @@ module.exports = function(source) {
   const config = this.exec(source, this.resourcePath);
   const plugins = getPluginList(config);
 
-  const exports = {
-    configs: [],
-    actions: [],
-    reducers: [],
-    components: [],
-    queries: [],
-    mutators: [],
-  };
-
+  const list = [];
   plugins.forEach(plugin => {
-    addIfExists(exports.components, plugin, 'index.js');
-    addIfExists(exports.configs, plugin, 'config.json');
-    addIfExists(exports.actions, plugin, 'actions');
-    addIfExists(exports.reducers, plugin, 'reducer');
-    addIfExists(exports.mutators, plugin, 'mutations');
-    addIfExists(exports.queries, plugin, 'queries');
+    list.push(`{module: require('plugins/${plugin}/client/index.js'), plugin: '${plugin}'}`);
   });
 
-  let result = '';
-
-  Object.keys(exports).forEach(key => {
-    if (result) { result += '\n\n'; }
-    result += stripIndent`
-      module.exports.${key} = [
-        ${exports[key].join(',')}
-      ];
-    `;
-  });
-
-  return result;
+  return stripIndent`
+    module.exports = [
+      ${list.join(',')}
+    ];
+  `;
 };
