+++ client/coral-framework/reducers/index.js
@@ -1,11 +1,11 @@
 import auth from './auth';
 import user from './user';
 import asset from './asset';
-import {getPluginReducers} from '../helpers/plugins';
+import {pluginReducers} from '../helpers/plugins';
 
 export default {
   auth,
   user,
   asset,
-  ...getPluginReducers()
+  ...pluginReducers
 };
