+++ plugins/coral-plugin-respect/client/components/style.css
