--- plugins/coral-plugin-respect/client/config.json
@@ -1,4 +1,0 @@
-{
-  "name": "Coral Plugin Respect",
-  "slot": "Comment.Detail"
-}
